requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-excel-response==2.0.5
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
et-xmlfile==1.1.0
frozenlist==1.4.0
idna==3.6
jdcal==1.4.1
multidict==6.0.4
openpyxl==2.4.11
Pillow==10.1.0
PyJWT==2.8.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.10.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
xlwt==1.3.0
yarl==1.9.3
